Add the bones of a Tic Tac Toe game
All of the game code is currently inside an iife called Game. I don't
know if this is good practice or not. I definitely like keeping some
objects private, but I wonder if I could accomplish the same thing with
multiple iife's or maybe a combination of iife's and factories.